+++ libraries/bio/src/utils/cell-renderer-monomer-placer.ts
@@ -12,14 +12,13 @@ import {ILogger} from './logger';
 import {getMonomerLibHelper} from '../monomer-works/monomer-utils';
 import {errInfo} from './err-info';
 import {DrawStyle, printLeftOrCentered, TAGS as mmcrTAGS, PrintOptions} from './cell-renderer';
-import {MmcrTemps, tempTAGS} from './cell-renderer-consts';
+import {MmcrTemps, rendererSettingsChangedState, tempTAGS} from './cell-renderer-consts';
 import {IMonomerLibBase} from '../types/index';
 import {HelmTypes} from '../helm/consts';
 import {ISeqMonomer} from '../helm/types';
 import {execMonomerHoverLinks} from '../monomer-works/monomer-hover';
 
 type MonomerPlacerProps = {
-  seqHandler: SeqHandler,
   monomerCharWidth: number, separatorWidth: number,
   monomerToShort: MonomerToShortFunc
 };
@@ -116,6 +115,7 @@ export class MonomerPlacer extends CellRendererBackBase<string> {
 
   /** Returns monomers lengths of the {@link rowIdx} and cumulative sums for borders, monomer places */
   public getCellMonomerLengths(rowIdx: number, newWidth: number): [number[], number[]] {
+    const sh = SeqHandler.forColumn(this.tableCol);
     if (this.colWidth < newWidth) {
       this.colWidth = newWidth;
       this.dirty = true;
@@ -127,7 +127,7 @@ export class MonomerPlacer extends CellRendererBackBase<string> {
       }
     }
 
-    const res: number[] = this.props.seqHandler.isMsa() ? this.getCellMonomerLengthsForSeqMsa() :
+    const res: number[] = sh.isMsa() ? this.getCellMonomerLengthsForSeqMsa() :
       this.getCellMonomerLengthsForSeq(rowIdx);
 
     const resSum: number[] = new Array<number>(res.length + 1);
@@ -144,9 +144,10 @@ export class MonomerPlacer extends CellRendererBackBase<string> {
     if (this._monomerLengthList === null)
       this._monomerLengthList = new Array(this.tableCol.length).fill(null);
 
+    const sh = SeqHandler.forColumn(this.tableCol);
     const minMonWidth = this.props.separatorWidth + 1 * this.props.monomerCharWidth;
     const maxVisibleSeqLength: number = Math.ceil(this.colWidth / minMonWidth);
-    const seqSS: ISeqSplitted = SeqHandler.forColumn(this.tableCol).getSplitted(rowIdx);
+    const seqSS: ISeqSplitted = sh.getSplitted(rowIdx);
     const visibleSeqLength: number = Math.min(maxVisibleSeqLength, seqSS.length);
 
     let res: number[] = this._monomerLengthList[rowIdx];
@@ -157,7 +158,7 @@ export class MonomerPlacer extends CellRendererBackBase<string> {
       for (let seqMonI = 0; seqMonI < visibleSeqLength; ++seqMonI) {
         const seqMonLabel = seqSS.getOriginal(seqMonI);
         const shortMon: string = this.props.monomerToShort(seqMonLabel, this.monomerLengthLimit);
-        const separatorWidth = this.props.seqHandler.isSeparator() ? this.separatorWidth : this.props.separatorWidth;
+        const separatorWidth = sh.isSeparator() ? this.separatorWidth : this.props.separatorWidth;
         const seqMonWidth: number = separatorWidth + shortMon.length * this.props.monomerCharWidth;
         res[seqMonI] = seqMonWidth;
         seqWidth += seqMonWidth;
@@ -241,10 +242,6 @@ export class MonomerPlacer extends CellRendererBackBase<string> {
     }
   }
 
-  public isMsa(): boolean {
-    return this.props.seqHandler.isMsa();
-  }
-
   private padding: number = 5;
 
   render(g: CanvasRenderingContext2D, x: number, y: number, w: number, h: number,
@@ -253,7 +250,6 @@ export class MonomerPlacer extends CellRendererBackBase<string> {
     const gridCol = this.gridCol;
     const tableCol = this.tableCol;
     const dpr = window.devicePixelRatio;
-    const separatorWidth = this.props.separatorWidth;
 
     // Cell renderer settings
     let maxLengthOfMonomer: number = this.monomerLengthLimit;
@@ -268,17 +264,21 @@ export class MonomerPlacer extends CellRendererBackBase<string> {
 
     g.save();
     try {
-      // if (
-      //   tableCol.temp[MmcrTemps.rendererSettingsChanged] === rendererSettingsChangedState.true ||
-      //   this.monomerLengthLimit != maxLengthOfMonomer
-      // ) {
-      //   gapLength = tableCol.temp[MmcrTemps.gapLength] as number ?? gapLength;
-      //   // this event means that the mm renderer settings have changed,
-      //   // particularly monomer representation and max width.
-      //   this.setMonomerLengthLimit(maxLengthOfMonomer);
-      //   this.setSeparatorWidth(this.isMsa() ? msaGapLength : gapLength);
-      //   tableCol.temp[MmcrTemps.rendererSettingsChanged] = rendererSettingsChangedState.false;
-      // }
+      const sh = SeqHandler.forColumn(tableCol);
+
+      if (
+        tableCol.temp[MmcrTemps.rendererSettingsChanged] === rendererSettingsChangedState.true ||
+        this.monomerLengthLimit != maxLengthOfMonomer
+      ) {
+        let gapLength = 0;
+        const msaGapLength = 8;
+        gapLength = tableCol.temp[MmcrTemps.gapLength] as number ?? gapLength;
+        // this event means that the mm renderer settings have changed,
+        // particularly monomer representation and max width.
+        this.setMonomerLengthLimit(maxLengthOfMonomer);
+        this.setSeparatorWidth(sh.isMsa() ? msaGapLength : gapLength);
+        tableCol.temp[MmcrTemps.rendererSettingsChanged] = rendererSettingsChangedState.false;
+      }
 
       const [maxLengthWords, maxLengthWordsSum]: [number[], number[]] =
         this.getCellMonomerLengths(gridCell.tableRowIndex!, w);
@@ -302,7 +302,6 @@ export class MonomerPlacer extends CellRendererBackBase<string> {
       const separator = tableCol.getTag(bioTAGS.separator) ?? '';
       const minMonWidth = this.props.separatorWidth + 1 * this.props.monomerCharWidth;
       const splitLimit = Math.ceil(w / minMonWidth);
-      const sh = SeqHandler.forColumn(tableCol);
 
       const tempReferenceSequence: string | null = tableCol.temp[tempTAGS.referenceSequence];
       const tempCurrentWord: string | null = this.tableCol.temp[tempTAGS.currentWord];
@@ -319,8 +318,6 @@ export class MonomerPlacer extends CellRendererBackBase<string> {
       })();
 
       const subParts: ISeqSplitted = sh.getSplitted(rowIdx);
-      /* let x1 = x; */
-      const color = undefinedColor;
       let drawStyle = DrawStyle.classic;
 
       if (aligned && aligned.includes('MSA') && units == NOTATION.SEPARATOR)
@@ -328,15 +325,14 @@ export class MonomerPlacer extends CellRendererBackBase<string> {
 
       const visibleSeqLength = Math.min(subParts.length, splitLimit);
       for (let posIdx: number = 0; posIdx < visibleSeqLength; ++posIdx) {
+        const om: string = subParts.getOriginal(posIdx);
         const cm: string = subParts.getCanonical(posIdx);
 
         let color = undefinedColor;
         if (this.monomerLib) {
-          const sh = SeqHandler.forColumn(this.tableCol);
           const biotype = sh.defaultBiotype;
           //this.logger.debug(`${logPrefix}, biotype: ${biotype}, amino: ${amino}`);
-          const wem = this.monomerLib.getWebEditorMonomer(biotype, cm)!;
-          color = wem.backgroundcolor!;
+          color = this.monomerLib.getMonomerTextColor(biotype, cm);
         }
         g.fillStyle = undefinedColor;
         const last = posIdx === subParts.length - 1;
@@ -347,7 +343,7 @@ export class MonomerPlacer extends CellRendererBackBase<string> {
           referenceSequence: referenceSequence, maxLengthOfMonomer: maxLengthOfMonomer,
           monomerTextSizeMap: this._monomerLengthMap, logger: this.logger
         };
-        printLeftOrCentered(g, cm, x + this.padding, y, w, h, opts);
+        printLeftOrCentered(g, om, x + this.padding, y, w, h, opts);
         if (minDistanceRenderer > w) break;
       }
     } catch (err: any) {